Refactor resources into modeule
